more oauth stuff ,basic membership service
